requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
elastic-transport==8.4.1
elasticsearch==6.8.1
frozenlist==1.4.0
idna==3.4
kafka-python==2.0.2
multidict==6.0.4
numpy==1.24.4
pandas==2.0.3
py4j==0.10.9.5
pyspark==3.3.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
urllib3==1.26.17
yarl==1.9.2
modAL-python @ git+https://github.com/modAL-python/modAL.git@bba6f6fd00dbb862b1e09259b78caf6cffa2e755
hyperopt==0.2.7
scikit-learn==1.3.2
mlflow==2.8.1